+++ Changes.rst
@@ -1,3 +1,12 @@
-Version 1.3.1 (2021-07-20)
+Version 1.1 (2021-07-20)
 --------------------------
 - Initial release.
+
+
+Version 1.2 (2021-08-03)
+--------------------------
+- Added index access in linked list implementation
+  - Example   L[i] returns the element at index "i" in the linked list. Where "i" is an integer.
+  
+- Added an insertion method which allow user to insert an element at a giving index to linked list
+  -Example if L is a linked list object, "L.insert(i, el)" will insert element 'el' at an index 'i'.